Refactor WSO2 event handler to the new folder structure
