+++ components/org.wso2.identity.webhook.wso2.event.handler/src/main/java/org/wso2/identity/webhook/wso2/event/handler/internal/model/WSO2AuthenticationFailedEventPayload.java
@@ -16,38 +16,41 @@
  * under the License.
  */
 
-package org.wso2.identity.webhook.wso2.event.handler.model;
+package org.wso2.identity.webhook.wso2.event.handler.internal.model;
 
-import org.wso2.identity.webhook.wso2.event.handler.model.common.Application;
-import org.wso2.identity.webhook.wso2.event.handler.model.common.Organization;
-import org.wso2.identity.webhook.wso2.event.handler.model.common.User;
-import org.wso2.identity.webhook.wso2.event.handler.model.common.UserStore;
-
-import java.util.ArrayList;
-import java.util.List;
+import org.wso2.identity.webhook.wso2.event.handler.internal.model.common.Application;
+import org.wso2.identity.webhook.wso2.event.handler.internal.model.common.Organization;
+import org.wso2.identity.webhook.wso2.event.handler.internal.model.common.User;
+import org.wso2.identity.webhook.wso2.event.handler.internal.model.common.UserStore;
 
 /**
- * Model Class for Authentication Success Event Payload.
+ * Model Class for Authentication Failed Event Payload.
  */
-public class WSO2AuthenticationSuccessEventPayload extends WSO2BaseEventPayload {
+public class WSO2AuthenticationFailedEventPayload extends WSO2BaseEventPayload {
 
-    private List<String> authenticationMethods = new ArrayList<>();
+    private final AuthenticationFailedReason reason;
+    private final String userLoginIdentifier;
 
-    private WSO2AuthenticationSuccessEventPayload(Builder builder) {
+    private WSO2AuthenticationFailedEventPayload(Builder builder) {
         this.user = builder.user;
         this.tenant = builder.tenant;
         this.userResidentOrganization = builder.userResidentOrganization;
         this.userStore = builder.userStore;
         this.application = builder.application;
-        this.authenticationMethods = builder.authenticationMethods;
+        this.reason = builder.reason;
+        this.userLoginIdentifier = builder.userLoginIdentifier;
     }
 
-    public List<String> getAuthenticationMethods() {
-        return authenticationMethods;
+    public AuthenticationFailedReason getReason() {
+        return reason;
+    }
+
+    public String getUserLoginIdentifier() {
+        return userLoginIdentifier;
     }
 
     /**
-     * Builder class to build WSO2AuthenticationSuccessEventPayload.
+     * Builder class to build WSO2AuthenticationFailedEventPayload.
      */
     public static class Builder {
         private User user;
@@ -55,7 +58,8 @@ public class WSO2AuthenticationSuccessEventPayload extends WSO2BaseEventPayload 
         private Organization userResidentOrganization;
         private UserStore userStore;
         private Application application;
-        private List<String> authenticationMethods = new ArrayList<>();
+        private AuthenticationFailedReason reason;
+        private String userLoginIdentifier;
 
         public Builder user(User user) {
             this.user = user;
@@ -82,18 +86,18 @@ public class WSO2AuthenticationSuccessEventPayload extends WSO2BaseEventPayload 
             return this;
         }
 
-        public Builder authenticationMethods(List<String> authenticationMethods) {
-            this.authenticationMethods = authenticationMethods;
+        public Builder reason(AuthenticationFailedReason reason) {
+            this.reason = reason;
             return this;
         }
 
-        public Builder addAuthenticationMethod(String authenticationMethod) {
-            this.authenticationMethods.add(authenticationMethod);
+        public Builder userLoginIdentifier(String userLoginIdentifier) {
+            this.userLoginIdentifier = userLoginIdentifier;
             return this;
         }
 
-        public WSO2AuthenticationSuccessEventPayload build() {
-            return new WSO2AuthenticationSuccessEventPayload(this);
+        public WSO2AuthenticationFailedEventPayload build() {
+            return new WSO2AuthenticationFailedEventPayload(this);
         }
     }
 }
